+++ pkg/manager/manager_test.go
@@ -34,10 +34,8 @@ import (
 	. "github.com/onsi/gomega"
 	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/goleak"
-	authenticationv1 "k8s.io/api/authentication/v1"
 	coordinationv1 "k8s.io/api/coordination/v1"
 	corev1 "k8s.io/api/core/v1"
-	rbacv1 "k8s.io/api/rbac/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -45,8 +43,6 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/leaderelection/resourcelock"
 	configv1alpha1 "k8s.io/component-base/config/v1alpha1"
-	"k8s.io/utils/pointer"
-
 	"sigs.k8s.io/controller-runtime/pkg/cache"
 	"sigs.k8s.io/controller-runtime/pkg/cache/informertest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -56,7 +52,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/leaderelection"
 	fakeleaderelection "sigs.k8s.io/controller-runtime/pkg/leaderelection/fake"
 	"sigs.k8s.io/controller-runtime/pkg/metrics"
-	"sigs.k8s.io/controller-runtime/pkg/metrics/filters"
 	metricsserver "sigs.k8s.io/controller-runtime/pkg/metrics/server"
 	"sigs.k8s.io/controller-runtime/pkg/recorder"
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
@@ -1401,245 +1396,6 @@ var _ = Describe("manger.Manager", func() {
 				Expect(string(body)).To(Equal("Some debug info"))
 			})
 		})
-
-		Context("should start serving metrics with https and authn/authz", func() {
-			var srv metricsserver.Server
-			var defaultServer metricsDefaultServer
-			var opts Options
-			var httpClient *http.Client
-
-			BeforeEach(func() {
-				srv = nil
-				opts = Options{
-					Metrics: metricsserver.Options{
-						BindAddress:    ":0",
-						SecureServing:  true,
-						FilterProvider: filters.WithAuthenticationAndAuthorization,
-					},
-					newMetricsServer: func(options metricsserver.Options, config *rest.Config, httpClient *http.Client) (metricsserver.Server, error) {
-						var err error
-						srv, err = metricsserver.NewServer(options, config, httpClient)
-						if srv != nil {
-							defaultServer = srv.(metricsDefaultServer)
-						}
-						return srv, err
-					},
-				}
-				httpClient = &http.Client{Transport: &http.Transport{
-					TLSClientConfig: &tls.Config{InsecureSkipVerify: true}, //nolint:gosec
-				}}
-			})
-
-			It("should stop serving metrics when stop is called", func() {
-				m, err := New(cfg, opts)
-				Expect(err).NotTo(HaveOccurred())
-
-				ctx, cancel := context.WithCancel(context.Background())
-				go func() {
-					defer GinkgoRecover()
-					Expect(m.Start(ctx)).NotTo(HaveOccurred())
-				}()
-				<-m.Elected()
-				// Note: Wait until metrics server has been started. A finished leader election
-				// doesn't guarantee that the metrics server is up.
-				Eventually(func() string { return defaultServer.GetBindAddr() }, 10*time.Second).ShouldNot(BeEmpty())
-
-				// Check the metrics started
-				// GET /metrics without token.
-				endpoint := fmt.Sprintf("https://%s/metrics", defaultServer.GetBindAddr())
-				resp, err := httpClient.Get(endpoint)
-				Expect(err).NotTo(HaveOccurred())
-				body, err := io.ReadAll(resp.Body)
-				Expect(resp.StatusCode).To(Equal(401))
-				Expect(err).NotTo(HaveOccurred())
-				// Unauthorized is the expected response if no bearer token is provided.
-				Expect(string(body)).To(ContainSubstring("Unauthorized"))
-
-				// Shutdown the server
-				cancel()
-
-				// Expect the metrics server to shutdown
-				Eventually(func() error {
-					_, err = http.Get(endpoint)
-					return err
-				}, 10*time.Second).ShouldNot(Succeed())
-			})
-
-			It("should serve metrics endpoint", func() {
-				m, err := New(cfg, opts)
-				Expect(err).NotTo(HaveOccurred())
-
-				ctx, cancel := context.WithCancel(context.Background())
-				defer cancel()
-				go func() {
-					defer GinkgoRecover()
-					Expect(m.Start(ctx)).NotTo(HaveOccurred())
-				}()
-				<-m.Elected()
-				// Note: Wait until metrics server has been started. A finished leader election
-				// doesn't guarantee that the metrics server is up.
-				Eventually(func() string { return defaultServer.GetBindAddr() }, 10*time.Second).ShouldNot(BeEmpty())
-
-				// Setup service account with rights to "/metrics"
-				token, cleanup, err := setupServiceAccountForURL(ctx, m.GetClient(), "/metrics")
-				defer cleanup()
-				Expect(err).ToNot(HaveOccurred())
-
-				// GET /metrics with token.
-				metricsEndpoint := fmt.Sprintf("https://%s/metrics", defaultServer.GetBindAddr())
-				req, err := http.NewRequest("GET", metricsEndpoint, nil)
-				Expect(err).NotTo(HaveOccurred())
-				req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", token))
-				resp, err := httpClient.Do(req)
-				Expect(err).NotTo(HaveOccurred())
-				defer resp.Body.Close()
-				// This is expected as the token has rights for /metrics.
-				Expect(resp.StatusCode).To(Equal(200))
-				body, err := io.ReadAll(resp.Body)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(string(body)).To(ContainSubstring("rest_client_requests_total"))
-			})
-
-			It("should not serve anything other than metrics endpoint by default", func() {
-				m, err := New(cfg, opts)
-				Expect(err).NotTo(HaveOccurred())
-
-				ctx, cancel := context.WithCancel(context.Background())
-				defer cancel()
-				go func() {
-					defer GinkgoRecover()
-					Expect(m.Start(ctx)).NotTo(HaveOccurred())
-				}()
-				<-m.Elected()
-				// Note: Wait until metrics server has been started. A finished leader election
-				// doesn't guarantee that the metrics server is up.
-				Eventually(func() string { return defaultServer.GetBindAddr() }, 10*time.Second).ShouldNot(BeEmpty())
-
-				endpoint := fmt.Sprintf("https://%s/should-not-exist", defaultServer.GetBindAddr())
-				resp, err := httpClient.Get(endpoint)
-				Expect(err).NotTo(HaveOccurred())
-				defer resp.Body.Close()
-				Expect(resp.StatusCode).To(Equal(404))
-			})
-
-			It("should serve metrics in its registry", func() {
-				one := prometheus.NewCounter(prometheus.CounterOpts{
-					Name: "test_one",
-					Help: "test metric for testing",
-				})
-				one.Inc()
-				err := metrics.Registry.Register(one)
-				Expect(err).NotTo(HaveOccurred())
-
-				m, err := New(cfg, opts)
-				Expect(err).NotTo(HaveOccurred())
-
-				ctx, cancel := context.WithCancel(context.Background())
-				defer cancel()
-				go func() {
-					defer GinkgoRecover()
-					Expect(m.Start(ctx)).NotTo(HaveOccurred())
-				}()
-				<-m.Elected()
-				// Note: Wait until metrics server has been started. A finished leader election
-				// doesn't guarantee that the metrics server is up.
-				Eventually(func() string { return defaultServer.GetBindAddr() }, 10*time.Second).ShouldNot(BeEmpty())
-
-				// Setup service account with rights to "/metrics"
-				token, cleanup, err := setupServiceAccountForURL(ctx, m.GetClient(), "/metrics")
-				defer cleanup()
-				Expect(err).ToNot(HaveOccurred())
-
-				// GET /metrics with token.
-				metricsEndpoint := fmt.Sprintf("https://%s/metrics", defaultServer.GetBindAddr())
-				req, err := http.NewRequest("GET", metricsEndpoint, nil)
-				Expect(err).NotTo(HaveOccurred())
-				req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", token))
-				resp, err := httpClient.Do(req)
-				Expect(err).NotTo(HaveOccurred())
-				defer resp.Body.Close()
-				// This is expected as the token has rights for /metrics.
-				Expect(resp.StatusCode).To(Equal(200))
-
-				data, err := io.ReadAll(resp.Body)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(string(data)).To(ContainSubstring("%s\n%s\n%s\n",
-					`# HELP test_one test metric for testing`,
-					`# TYPE test_one counter`,
-					`test_one 1`,
-				))
-
-				// Unregister will return false if the metric was never registered
-				ok := metrics.Registry.Unregister(one)
-				Expect(ok).To(BeTrue())
-			})
-
-			It("should serve extra endpoints", func() {
-				opts.Metrics.ExtraHandlers = map[string]http.Handler{
-					"/debug": http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-						_, _ = w.Write([]byte("Some debug info"))
-					}),
-				}
-				m, err := New(cfg, opts)
-				Expect(err).NotTo(HaveOccurred())
-
-				ctx, cancel := context.WithCancel(context.Background())
-				defer cancel()
-				go func() {
-					defer GinkgoRecover()
-					Expect(m.Start(ctx)).NotTo(HaveOccurred())
-				}()
-				<-m.Elected()
-				// Note: Wait until metrics server has been started. A finished leader election
-				// doesn't guarantee that the metrics server is up.
-				Eventually(func() string { return defaultServer.GetBindAddr() }, 10*time.Second).ShouldNot(BeEmpty())
-
-				// Setup service account with rights to "/debug"
-				token, cleanup, err := setupServiceAccountForURL(ctx, m.GetClient(), "/debug")
-				defer cleanup()
-				Expect(err).ToNot(HaveOccurred())
-
-				// GET /debug without token.
-				endpoint := fmt.Sprintf("https://%s/debug", defaultServer.GetBindAddr())
-				req, err := http.NewRequest("GET", endpoint, nil)
-				Expect(err).NotTo(HaveOccurred())
-				resp, err := httpClient.Do(req)
-				Expect(err).NotTo(HaveOccurred())
-				defer resp.Body.Close()
-				// This is expected as we didn't send a token.
-				Expect(resp.StatusCode).To(Equal(401))
-				body, err := io.ReadAll(resp.Body)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(string(body)).To(ContainSubstring("Unauthorized"))
-
-				// PUT /debug with token.
-				req, err = http.NewRequest("PUT", endpoint, nil)
-				Expect(err).NotTo(HaveOccurred())
-				req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", token))
-				resp, err = httpClient.Do(req)
-				Expect(err).NotTo(HaveOccurred())
-				defer resp.Body.Close()
-				// This is expected as the token has rights for /debug.
-				Expect(resp.StatusCode).To(Equal(200))
-				body, err = io.ReadAll(resp.Body)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(string(body)).To(Equal("Some debug info"))
-
-				// GET /metrics with token (but token only has rights for /debug).
-				metricsEndpoint := fmt.Sprintf("https://%s/metrics", defaultServer.GetBindAddr())
-				req, err = http.NewRequest("GET", metricsEndpoint, nil)
-				Expect(err).NotTo(HaveOccurred())
-				req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", token))
-				resp, err = httpClient.Do(req)
-				Expect(err).NotTo(HaveOccurred())
-				defer resp.Body.Close()
-				Expect(resp.StatusCode).To(Equal(403))
-				body, err = io.ReadAll(resp.Body)
-				Expect(err).NotTo(HaveOccurred())
-				// Authorization denied is expected as the token only has rights for /debug not for /metrics.
-				Expect(string(body)).To(ContainSubstring("Authorization denied for user system:serviceaccount:default:metrics-test"))
-			})
-		})
 	})
 
 	Context("should start serving health probes", func() {
@@ -2092,79 +1848,6 @@ var _ = Describe("manger.Manager", func() {
 		Expect(m.GetAPIReader()).NotTo(BeNil())
 	})
 })
-
-func setupServiceAccountForURL(ctx context.Context, c client.Client, path string) (string, func(), error) {
-	createdObjects := []client.Object{}
-	cleanup := func() {
-		for _, obj := range createdObjects {
-			_ = c.Delete(ctx, obj)
-		}
-	}
-
-	sa := &corev1.ServiceAccount{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "metrics-test",
-			Namespace: metav1.NamespaceDefault,
-		},
-	}
-	if err := c.Create(ctx, sa); err != nil {
-		return "", cleanup, err
-	}
-	createdObjects = append(createdObjects, sa)
-
-	cr := &rbacv1.ClusterRole{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "metrics-test",
-		},
-		Rules: []rbacv1.PolicyRule{
-			{
-				Verbs:           []string{"get", "put"},
-				NonResourceURLs: []string{path},
-			},
-		},
-	}
-	if err := c.Create(ctx, cr); err != nil {
-		return "", cleanup, err
-	}
-	createdObjects = append(createdObjects, cr)
-
-	crb := &rbacv1.ClusterRoleBinding{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "metrics-test",
-		},
-		Subjects: []rbacv1.Subject{
-			{
-				Kind:      rbacv1.ServiceAccountKind,
-				Name:      "metrics-test",
-				Namespace: metav1.NamespaceDefault,
-			},
-		},
-		RoleRef: rbacv1.RoleRef{
-			APIGroup: rbacv1.GroupName,
-			Kind:     "ClusterRole",
-			Name:     "metrics-test",
-		},
-	}
-	if err := c.Create(ctx, crb); err != nil {
-		return "", cleanup, err
-	}
-	createdObjects = append(createdObjects, crb)
-
-	tokenRequest := &authenticationv1.TokenRequest{
-		Spec: authenticationv1.TokenRequestSpec{
-			ExpirationSeconds: pointer.Int64(2 * 60 * 60), // 2 hours.
-		},
-	}
-	if err := c.SubResource("token").Create(ctx, sa, tokenRequest); err != nil {
-		return "", cleanup, err
-	}
-
-	if tokenRequest.Status.Token == "" {
-		return "", cleanup, errors.New("failed to get ServiceAccount token: token should not be empty")
-	}
-
-	return tokenRequest.Status.Token, cleanup, nil
-}
 
 type runnableError struct {
 }
